+++ 臺灣言語工具/羅馬字/台語/__init__.py
@@ -1,4 +1,5 @@
-from 臺灣言語工具.音標系統.閩南語.教會羅馬字音標 import 教會羅馬字音標 as 白話字
-from 臺灣言語工具.音標系統.閩南語.臺灣閩南語羅馬字拼音相容教會羅馬字音標 import 臺灣閩南語羅馬字拼音相容教會羅馬字音標 as 新白話字
+from 臺灣言語工具.羅馬字.台語.POJ import 白話字
+from 臺灣言語工具.羅馬字.台語.SPOJ import 新白話字
+
 
 __all__ = ['白話字', '新白話字', ]
